Split match analyzer into multiple partials
